chore: restructure script and moved code in seperate files
