+++ script/src/proof/trie.rs
